# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College LMS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: semester
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # FACULTY
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: semester
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}/attempt:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "201":
          description: Attempt submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttempt"
  /quizzes/{id}/attempts:
    get:
      operationId: getQuizAttempts
      tags: [quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizAttempt"

  # RESULTS
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List results
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: semester
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: createResult
      tags: [results]
      summary: Create/upload result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # STUDY MATERIALS
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, email, rollNumber, semester, departmentId]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        rollNumber:
          type: string
        phone:
          type: ["string", "null"]
        semester:
          type: integer
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        enrolledAt:
          type: string
    StudentInput:
      type: object
      required: [name, email, rollNumber, semester, departmentId, password]
      properties:
        name:
          type: string
        email:
          type: string
        rollNumber:
          type: string
        phone:
          type: ["string", "null"]
        semester:
          type: integer
        departmentId:
          type: integer
        password:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        semester:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
    StudentListResponse:
      type: object
      required: [students, total, page, limit]
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Quiz:
      type: object
      required: [id, title, subjectId, facultyId, durationMinutes, totalMarks]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        facultyId:
          type: integer
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        semester:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizQuestion:
      type: object
      required: [id, question, options, correctOption]
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        marks:
          type: integer
    QuizInput:
      type: object
      required: [title, subjectId, facultyId, durationMinutes, totalMarks, questions]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: integer
        facultyId:
          type: integer
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        semester:
          type: ["integer", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionInput"
    QuizQuestionInput:
      type: object
      required: [question, options, correctOption]
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        marks:
          type: integer
    QuizAttempt:
      type: object
      required: [id, quizId, studentId, score, submittedAt]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        score:
          type: integer
        totalMarks:
          type: integer
        submittedAt:
          type: string
        answers:
          type: array
          items:
            type: integer
    QuizAttemptInput:
      type: object
      required: [studentId, answers]
      properties:
        studentId:
          type: integer
        answers:
          type: array
          items:
            type: integer
    Result:
      type: object
      required: [id, studentId, subjectId, semester, internalMarks, externalMarks, totalMarks, grade]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        semester:
          type: integer
        internalMarks:
          type: number
        externalMarks:
          type: number
        totalMarks:
          type: number
        grade:
          type: string
        passed:
          type: boolean
    ResultInput:
      type: object
      required: [studentId, subjectId, semester, internalMarks, externalMarks]
      properties:
        studentId:
          type: integer
        subjectId:
          type: integer
        semester:
          type: integer
        internalMarks:
          type: number
        externalMarks:
          type: number
...